+++ src/main/resources/assets/appliedenergistics2/lang/en_us.json
@@ -712,7 +712,7 @@
   "jei.appliedenergistics2.requires_processing_mode": "Requires processing mode",
   "key.appliedenergistics2.category": "Applied Energistics 2",
   "key.toggle_focus.desc": "Toggle search box focus",
-  "rei.appliedenergistics2.crystal_growing_category": "Growing Crystals",
+  "rei.appliedenergistics2.throwing_in_water_category": "Throwing In Water",
   "rei.appliedenergistics2.with_crystal_growth_accelerators": "With Crystal Growth Accelerators:",
   "stat.appliedenergistics2.items_extracted": "Items extracted from ME Storage",
   "stat.appliedenergistics2.items_inserted": "Items added to ME Storage",
